backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
google-cloud-texttospeech>=2.14.0
python-dotenv>=1.0.0
